requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-openai
pdfminer.six
pdf2image
pytesseract
faiss-cpu
PyYAML
gradio
keyring
pypdf

openai-whisper


openai-whisper

language_tool_python


